bot/pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Tg bot for Manager App"
readme = "README.md"
authors = [{ name = "Tati Trif", email = "yatatitrif@yandex.com" }]
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=1.2.1",
]
